+++ src/undo-scripts/undo-lab-aws-cloud-infra.ps1
@@ -1,9 +1,6 @@
 param(
-    [string[]] $workshopCfnStacks = @(
-        "Unicorn-Store-CI-CD-PipelineStack",
-        "UnicornSuperstoreStack"
-    ),
-    [string] $ecrRepoName = "unicorn-store-app"
+    [string] $workshopCfnStacks,
+    [string] $ecrRepoName
 )
 
 Import-Module awspowershell.netcore # ToDo: Import into system's profile
@@ -12,13 +9,21 @@ Push-Location
 Set-Location ([System.IO.Path]::GetDirectoryName($myInvocation.MyCommand.Path))
 # Include scripts with utility functions
 . ../aws.ps1
+. ../sys.ps1
 Pop-Location
 
-ConfigureCurrentAwsRegion -profileName $null
-DeleteCfnStacks($workshopCfnStacks)
+$workshopCfnStacks = CoalesceWithEnvVar $workshopCfnStacks "UNICORN_LAB_AWS_RIP_CFNS"
+[string[]] $cfnStacksToDelete = $workshopCfnStacks.Split(",")
 
-Remove-ECRRepository -RepositoryName $ecrRepoName -IgnoreExistingImages $true -Force -ErrorAction SilentlyContinue
-Write-Information "Removed ECR repository `"$ecrRepoName`""
+ConfigureCurrentAwsRegion -profileName $null
+DeleteCfnStacks($cfnStacksToDelete)
+
+$ecrRepoName = CoalesceWithEnvVar $ecrRepoName "UNICORN_LAB_AWS_RIP_ECR"
+if($ecrRepoName)
+{
+    Remove-ECRRepository -RepositoryName $ecrRepoName -IgnoreExistingImages $true -Force -ErrorAction SilentlyContinue
+    Write-Information "Removed ECR repository `"$ecrRepoName`""
+}
 
 TerminateInstanceByName(MakeLinuxInstanceName)
 SetDockerHostUserEnvVar # Removes "DOCKER_HOST" env var
